notifications_api/requirements.txt
Django==1.5.4
djangorestframework==2.3.8
wsgiref==0.1.2
